+++ js/dev/ajax/fetchAPIs.js
@@ -1,0 +1,74 @@
+/*-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-
+  build query url and request AJAX
+-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*/
+
+export function requestWeather(loc){
+// building the uri weather query string
+  var url = "http://api.openweathermap.org/data/2.5/weather?";
+
+  var key = "&units=imperial&APPID=79aef489883f75aff91f8900796eb1ea";
+
+  var lat = "lat=" + loc.lat.toString().slice(0,loc.lat.toString().indexOf(".")+3);
+
+  var lon = "lon=" + loc.lon.toString().slice(0,loc.lon.toString().indexOf(".")+3);
+
+  url += lat + "&" + lon + key;
+
+  //console.log("url =", url);
+
+/*-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-
+  AJAX weather request
+-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*/
+  $.ajax({
+    url: url,
+    type: 'GET',
+    dataType: 'jsonp',
+
+  /*-_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*
+    success function
+  -_-*-_-*-_-*-_-*-_-*-_-*-_-*-_-*/
+    success: function publishWeather(api){
+      //console.log("weather object: ", api);
+      //console.log("api.main.temp = ", api.main.temp);
+      temperature = Math.round(api.main.temp);
+      //console.log("rounded 'temperature' = ", temperature);
+      $(".temp").text(temperature + "\xb0 F").addClass("degF");
+      $(".weather").text(api.weather[0].description);
+      animateWeather(api.weather[0].id);
+    },
+    xhrField: {
+      withCredentials: true
+    },
+    error: function(api,errorText) {
+      //console.log(api);
+      $(".temp").text("There was a problem with the weather request<br>" + errorText);
+    }
+  });
+
+}
+
+/*-_-*-_-*-_-*-_-*-_-*-_-*
+ function to get location
+ then fire location-dependent fcns
+ such as get weather fcn
+_-*-_-*-_-*-_-*-_-*-_-*-_-*/
+export default function getLocalWeather(locURL){
+  $.ajax({
+    url: locURL,
+    type: 'GET',
+    dataType: 'jsonp',
+
+    success: function(json){
+      //console.log("location object =", json);
+      requestWeather(json);
+      $(".locale").text(json.city + ", " + json.regionName + " " + json.zip + ", " + json.country);
+    },
+    xhrField: {
+      withCredentials: true
+    },
+    error: function(json,errorText) {
+      //console.log(json);
+      $(".locale").text("There was a problem with the location request: ", errorText);
+    }
+  });
+}
