+++ js/dev/utils/randomRolls.js
@@ -1,0 +1,7 @@
+export default function randomInt(min, max) {
+  return Math.floor(Math.random * (max - min + 1) + min);
+}
+
+export function randomFlt(min, max) {
+  return Math.random * (max - min + 1) + min;
+}
